spi: Transform the FSL QuadSPI driver to use the SPI MEM API
To support the SPI MEM API, instead of modifying the existing U-Boot
driver, this patch adds a port of the existing Linux driver.
This also has the advantage that porting changes and fixes from Linux
will be easier.

So far this has been tested by reading and writing from and to a SPI
NAND chip.

Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>